separated training scripts
